\name{EDMeta-package}
\alias{EDMeta-package}
\alias{EDMeta}
\docType{package}
\title{
\packageTitle{EDMeta}
}
\description{
\packageDescription{EDMeta}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{EDMeta}

Maintainer: \packageMaintainer{EDMeta}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = EDMeta)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{EDMeta}}
%\section{Documentation Index}{\packageIndices{EDMeta}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
