\name{ERMeta-package}
\alias{ERMeta-package}
\alias{ERMeta}
\docType{package}
\title{
\packageTitle{ERMeta}
}
\description{
\packageDescription{ERMeta}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{ERMeta}

Maintainer: \packageMaintainer{ERMeta}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = ERMeta)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{ERMeta}}
%\section{Documentation Index}{\packageIndices{ERMeta}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
